fix: consolidate runtime pause test helpers for pkg-file-count
